Various changes including ability to create new databases
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk@297 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ pwsafe/Java/PasswordSafeLib/src/org/pwsafe/lib/file/PwsIntegerField.java
@@ -44,7 +44,7 @@ public class PwsIntegerField extends PwsField
 		int		value;
 		byte	retval[];
 
-		value	= (int) ((Integer) super.getValue()).intValue();
+		value	= ((Integer) super.getValue()).intValue();
 		retval	= PwsFile.allocateBuffer( 4 );
 
 		Util.putIntToByteArray( retval, value, 0 );
